requirements.txt

streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
segment-anything==1.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
scikit-image==0.22.0
matplotlib==3.8.3
albumentations==1.3.1
timm==0.9.16